Some fixes and things

+++ src/main/java/dmr/DragonMounts/registry/ModComponents.java
@@ -26,12 +26,12 @@ public class ModComponents {
     public static final Supplier<DataComponentType<String>> EGG_OWNER =
             register("egg_owner", Codec.STRING, ByteBufCodecs.STRING_UTF8);
 
-    public static final Supplier<DataComponentType<Float>> DRAGON_HEALTH_ATTRIBUTE =
-            register("dragon_health_attribute", Codec.FLOAT, ByteBufCodecs.FLOAT);
-    public static final Supplier<DataComponentType<Float>> DRAGON_MOVEMENT_SPEED_ATTRIBUTE =
-            register("dragon_movement_speed_attribute", Codec.FLOAT, ByteBufCodecs.FLOAT);
-    public static final Supplier<DataComponentType<Float>> DRAGON_ATTACK_ATTRIBUTE =
-            register("dragon_attack_attribute", Codec.FLOAT, ByteBufCodecs.FLOAT);
+    public static final Supplier<DataComponentType<Double>> DRAGON_HEALTH_ATTRIBUTE =
+            register("dragon_health_attribute", Codec.DOUBLE, ByteBufCodecs.DOUBLE);
+    public static final Supplier<DataComponentType<Double>> DRAGON_MOVEMENT_SPEED_ATTRIBUTE =
+            register("dragon_movement_speed_attribute", Codec.DOUBLE, ByteBufCodecs.DOUBLE);
+    public static final Supplier<DataComponentType<Double>> DRAGON_ATTACK_ATTRIBUTE =
+            register("dragon_attack_attribute", Codec.DOUBLE, ByteBufCodecs.DOUBLE);
 
     private static <T> Supplier<DataComponentType<T>> register(
             String name, Codec<T> codec, StreamCodec<? super RegistryFriendlyByteBuf, T> streamCodec) {
